Added a Debian package recipe.
There's a lot of weird files with silly contents and excessively strict
formatting rules, but last time I checked, all those are unfortunately
needed (and there isn't anything extra either, except a .gitignore
and a README).

The README contains basic instructions as well as a suggested workflow
when tagging the next release.